Add files via upload
This program will first create a MongoDB Atlas client and connect to the database. Then, it will get the data from the database and create a TensorFlow dataset. Next, it will create a TensorFlow model and train it on the data. Finally, it will make a prediction and print it out.
The sgd optimizer is a stochastic gradient descent algorithm. It is a simple and efficient algorithm that can be used to train neural networks. In this example, we are using the sgd optimizer with a learning rate of 0.01. The learning rate is a hyperparameter that controls how quickly the model learns. A smaller learning rate will cause the model to learn more slowly, but it will also be more stable. A larger learning rate will cause the model to learn more quickly, but it may be less stable.

We can also use a smaller batch size to make the program more efficient. The batch size is the number of examples that are used to update the model weights in each iteration. A smaller batch size will require less memory and will be faster to train. However, a smaller batch size may also be less accurate.